fix(character-data): CharacterRegenData 이동 관련 타입 오류 수정
한글 설명:
- `CharacterRegenData`의 `MoveStep`, `MoveSpeed` 타입을 `int`에서 `float`으로 수정
- 이동 값 정밀도가 손실되던 문제 해결
- 실제 이동/보간 로직과 데이터 타입 일관성 확보

English description:
- Change `MoveStep` and `MoveSpeed` types in `CharacterRegenData` from `int` to `float`
- Fix precision loss issues in movement-related values
- Ensure consistency between data definitions and movement/interpolation logic

+++ Runtime/Characters/CharacterRegenData.cs
@@ -10,12 +10,12 @@ namespace GGemCo2DCore
         public float x, y, z;
         public bool IsFlip;
         public bool DefaultVisible;
-        public int MoveStep;
-        public int MoveSpeed;
+        public float MoveStep;
+        public float MoveSpeed;
         public bool CanMoveX;
         public bool CanMoveY;
 
-        public CharacterRegenData(int uid, Vector3 position, bool flip, int mapUid, bool defaultVisible, int moveStep = 0, int moveSpeed = 0, bool canMoveX = true, bool canMoveY = true)
+        public CharacterRegenData(int uid, Vector3 position, bool flip, int mapUid, bool defaultVisible, float moveStep = 0, float moveSpeed = 0, bool canMoveX = true, bool canMoveY = true)
         {
             Uid = uid;
             MapUid = mapUid;
